refactor(web): remove orphaned Settings schedules page and dead deps
The Settings schedules page became unreachable when schedules moved to the
Activity page (no router route or import referenced it). Remove it along with
the components only it used — ScheduleDetailView, ScheduleRow,
SystemTaskDetailView (its home usage was replaced by SystemTaskDetailPanel in
#35267), and ScheduleListColumnsHeader — plus the dead routes.settings.schedule
URL helpers. Preserve live coverage by pruning the misnamed schedules-page.test
(which tested components, not the page) down to its still-relevant cases and
relocating it to components/schedule-components.test.ts.
