--- clients/web/src/domains/settings/components/system-task-detail-view.tsx
@@ -1,230 +1,0 @@
-import { useInfiniteQuery } from "@tanstack/react-query";
-import { ArrowLeft, Loader2, Play, Settings } from "lucide-react";
-
-import { DetailCard } from "@/components/detail-card";
-import { SCHEDULE_RUNS_PAGE_SIZE } from "@/domains/settings/api/schedules";
-import {
-  ModelProfileRow,
-  type ScheduleModelProfileCallSite,
-} from "@/domains/settings/components/model-profile-row";
-import { RecentRunsCard } from "@/domains/settings/components/recent-runs-card";
-import {
-  flattenRunPages,
-  formatTimestamp,
-} from "@/domains/settings/utils/schedule-formatters";
-import { toScheduleRun } from "@/domains/settings/utils/system-task-run-transforms";
-import {
-  consolidationRunsGetInfiniteOptions,
-  heartbeatRunsGetInfiniteOptions,
-  retrospectiveRunsGetInfiniteOptions,
-} from "@/generated/daemon/@tanstack/react-query.gen";
-import { Button } from "@vellumai/design-library/components/button";
-import { Notice } from "@vellumai/design-library/components/notice";
-import { Tag } from "@vellumai/design-library/components/tag";
-import { Toggle } from "@vellumai/design-library/components/toggle";
-
-import type { SystemTaskKind } from "@/domains/settings/types/schedules";
-
-const SYSTEM_TASK_PROFILE_CALL_SITES: Record<
-  SystemTaskKind,
-  ScheduleModelProfileCallSite
-> = {
-  heartbeat: "heartbeatAgent",
-  consolidation: "memoryV2Consolidation",
-  retrospective: "memoryRetrospective",
-};
-
-interface SystemTaskDetailViewProps {
-  kind: SystemTaskKind;
-  assistantId: string;
-  name: string;
-  subtitle: string;
-  enabled: boolean;
-  nextRunAt: number | null;
-  lastRunAt: number | null;
-  isRunning: boolean;
-  onBack: () => void;
-  /**
-   * Triggers an immediate run. Omit for event-driven tasks (memory
-   * retrospective) that have nothing global to trigger — the Run now
-   * button is hidden when absent.
-   */
-  onRunNow?: () => void;
-  /** Pauses/resumes automatic runs. Manual Run now stays available. */
-  onToggleEnabled?: (enabled: boolean) => void;
-  onOpenMemorySettings?: () => void;
-}
-
-export function SystemTaskDetailView({
-  kind,
-  assistantId,
-  name,
-  subtitle,
-  enabled,
-  nextRunAt,
-  lastRunAt,
-  isRunning,
-  onBack,
-  onRunNow,
-  onToggleEnabled,
-  onOpenMemorySettings,
-}: SystemTaskDetailViewProps) {
-  // Consolidation and retrospective are both owned by Memory: no toggle of
-  // their own, paused when Memory is off. Retrospective additionally has no
-  // global schedule (event-driven per conversation), so it hides Next run
-  // and is rendered without an onRunNow handler.
-  const isMemoryManaged = kind !== "heartbeat";
-  const isRetrospective = kind === "retrospective";
-  const isMemoryPaused = isMemoryManaged && !enabled;
-  const runNowDisabled = isRunning || isMemoryPaused;
-  const statusValue = isMemoryManaged
-    ? enabled
-      ? "On · Managed by Memory"
-      : "Paused"
-    : enabled
-      ? "Enabled"
-      : "Disabled";
-  const pausedNotice = isRetrospective
-    ? "Memory is off, so retrospectives are paused. Turn Memory back on to resume them."
-    : "Memory is off, so consolidation is paused. Turn Memory back on to resume consolidation.";
-
-  const opts = {
-    path: { assistant_id: assistantId },
-    query: { limit: SCHEDULE_RUNS_PAGE_SIZE },
-  };
-
-  // Extract queryKey/queryFn individually — spreading the full options union
-  // triggers TS2769 because the three generated option types have subtly
-  // different `enabled` callback generics that don't unify.
-  const infiniteOpts =
-    kind === "heartbeat"
-      ? heartbeatRunsGetInfiniteOptions(opts)
-      : kind === "consolidation"
-        ? consolidationRunsGetInfiniteOptions(opts)
-        : retrospectiveRunsGetInfiniteOptions(opts);
-
-  const { data, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage } =
-    useInfiniteQuery({
-      queryKey: infiniteOpts.queryKey,
-      queryFn: infiniteOpts.queryFn,
-      // Object page param — the generated queryFn merges it into the base
-      // query without adding a `before` cursor, fetching the most recent runs.
-      initialPageParam: { path: opts.path, query: {} },
-      getNextPageParam: (lastPage) => lastPage.nextCursor ?? undefined,
-      staleTime: 10_000,
-    });
-  const runs = flattenRunPages(
-    data?.pages.map((page) => ({
-      runs: page.runs.map((run) => toScheduleRun(run, kind)),
-    })),
-  );
-
-  return (
-    <div className="space-y-4">
-      <button
-        type="button"
-        onClick={onBack}
-        className="flex items-center gap-1.5 text-body-medium-lighter text-[var(--content-secondary)] hover:text-[var(--content-default)] transition-colors"
-      >
-        <ArrowLeft className="h-4 w-4" />
-        Back to schedules
-      </button>
-
-      <DetailCard
-        title={name}
-        subtitle={subtitle}
-        accessory={
-          <div className="flex items-center gap-2">
-            <Tag tone="neutral">system</Tag>
-            {isMemoryManaged && enabled && onOpenMemorySettings ? (
-              <Button
-                variant="outlined"
-                size="compact"
-                leftIcon={<Settings className="h-3.5 w-3.5" />}
-                onClick={onOpenMemorySettings}
-              >
-                Memory settings
-              </Button>
-            ) : null}
-            {onRunNow ? (
-              <Button
-                variant="outlined"
-                size="compact"
-                leftIcon={
-                  isRunning ? (
-                    <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                  ) : (
-                    <Play className="h-3.5 w-3.5" />
-                  )
-                }
-                onClick={onRunNow}
-                disabled={runNowDisabled}
-              >
-                {isRunning ? "Running…" : "Run now"}
-              </Button>
-            ) : null}
-          </div>
-        }
-      >
-        <div className="space-y-2 text-body-medium-lighter">
-          <ModelProfileRow
-            assistantId={assistantId}
-            defaultCallSite={SYSTEM_TASK_PROFILE_CALL_SITES[kind]}
-            fallbackLabel="Default (system task model)"
-            respectCallSiteOverride
-          />
-          <div className="flex items-center justify-between">
-            <span className="text-[var(--content-secondary)]">Status</span>
-            <span className="flex items-center gap-2">
-              <span>{statusValue}</span>
-              {!isMemoryManaged && onToggleEnabled ? (
-                <Toggle
-                  checked={enabled}
-                  onChange={onToggleEnabled}
-                  aria-label={`Toggle ${name}`}
-                />
-              ) : null}
-            </span>
-          </div>
-          {!isRetrospective ? (
-            <div className="flex items-center justify-between">
-              <span className="text-[var(--content-secondary)]">Next run</span>
-              <span>{formatTimestamp(nextRunAt)}</span>
-            </div>
-          ) : null}
-          <div className="flex items-center justify-between">
-            <span className="text-[var(--content-secondary)]">Last run</span>
-            <span>{formatTimestamp(lastRunAt)}</span>
-          </div>
-        </div>
-        {isMemoryPaused ? (
-          <Notice
-            tone="warning"
-            className="mt-4"
-            actions={
-              onOpenMemorySettings ? (
-                <Button
-                  variant="outlined"
-                  size="compact"
-                  onClick={onOpenMemorySettings}
-                >
-                  Turn on Memory
-                </Button>
-              ) : undefined
-            }
-          >
-            {pausedNotice}
-          </Notice>
-        ) : null}
-      </DetailCard>
-
-      <RecentRunsCard
-        runs={runs}
-        isLoading={isLoading}
-        hasMore={hasNextPage}
-        isLoadingMore={isFetchingNextPage}
-        onLoadMore={() => void fetchNextPage()}
-      />
-    </div>
-  );
-}
